bigbear-code-generator/pyproject.toml
[project]
name = "admin-platform"
version = "1.0.0"
description = "基于 FastAPI + Vue3 + Naive UI 的现代化轻量管理平台"
authors = [
    {name = "Admin Platform Team"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "starlette>=0.40.0",
    "sqlmodel>=0.0.16",
    "PyMySQL==1.1.0",
    "aiomysql==0.2.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "pydantic-settings==2.1.0",
    "pyyaml==6.0.1",
    "python-dateutil==2.8.2",
    "aiofiles==23.2.1",
    "jinja2==3.1.2",
    "typer==0.9.0",
    "minio==7.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
